+++ rollup.config.js
@@ -3,12 +3,22 @@ import buble from "rollup-plugin-buble";
 import closure from "rollup-plugin-closure-compiler-js";
 import uglify from "rollup-plugin-uglify-es";
 import pkg from "./package.json";
+import copy from "rollup-plugin-copy";
 
 const BUNDLE_NAME = "FastestValidator";
 
 // transpile ES2015+ to ES5
 const bublePlugin = buble({
 	exclude: ["node_modules/**", "examples/**", "dist/**", "test/**", "benchmark/**"]
+});
+
+// copy typescript definitions
+const copyPlugin = copy({
+	targets: [
+		{ src: "index.d.ts", dest: "dist" },
+	],
+	verbose: true,
+	copyOnce: true,
 });
 
 const bundles = [
@@ -24,7 +34,8 @@ const bundles = [
 		plugins: [
 			commonjs(),
 
-			bublePlugin
+			bublePlugin,
+			copyPlugin
 		]
 	},
 
@@ -52,6 +63,7 @@ const bundles = [
 			}),
 
 			bublePlugin,
+			copyPlugin,
 
 			uglify()
 		]